refactor(factories): consolidate noop implementations in factories/noop module
Move _create_noop_logger, _create_noop_metadata_builder, and
_create_noop_validator_factory from container.py to factories/noop.py.
Functions are now public (no underscore prefix) and exported via
factories/__init__.py.
